+++ core/src/main/java/com/homeaway/streamplatform/streamregistry/service/impl/StreamServiceImpl.java
@@ -126,7 +126,6 @@ public class StreamServiceImpl extends AbstractService implements StreamService 
 
         verifyAndUpsertTopics(stream, isNewStream);
         streamDao.upsertStream(key, avroStream);
-        log.info("Stream upserted for {}", stream.getName());
     }
 
     /**