+++ core/src/main/java/com/homeaway/streamplatform/streamregistry/utils/StreamRegistryUtils.java
@@ -98,4 +98,5 @@ public final class StreamRegistryUtils {
                 .map(Actor::getName)
                 .filter(actorName -> actorName.equalsIgnoreCase(name));
     }
+
 }